iio: adc: ads1015: use single lock for all synchronization
There should not be any operation on device once the device
is shutdown. To achieve this, the shutdown flag is set after
shutdown and check on start of channel read. The flag is
protected on different lock then iio_dev lock and there is
possibility to have shutdown during the channel read ie.e. adc
conversion. To avoid such case, use the same lock for channel
use and shutdown flag so that there will not be any race.

bug 200100664

Change-Id: I25bddfb030b13af5469191ca69c418f5de702fd4
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/744958