gpu: nvgpu: Fix OOM case for buddy allocator
The allocator could attempt to use a buddy list for an order larger than
the max order when all the valid buddy lists were empty. This patch
ensures that when looking for a particular order buddy that the passed
order is valid.

Also handle the prints in the no-mem case a litle differently. Only print
and update alloc info when there was a successful allocation.

Lastly print hex numbers for the allocator stats printing function.

Change-Id: If289f3e8925e236e3b7d84206a75bd45a14082a1
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/745071
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>