drivers: video: tegra: Fix HDMI_CTRL programming
This change fixes the programming for the NV_SOR_HDMI_CTRL programming
in the hdmi2.0 driver. The driver was not calculating the value of the
MAX_AC_PACKET correctly which should have been actually based on the
value of REKEY.

This change also cleans up the programming of this register and makes it
much more easier to read.

Bug 1648668

Change-Id: I9d4de4eec4a088719852ecec82c2a55de811c54d
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/745245
(cherry picked from commit c2ce93f123232e966b4b53367ee223ae5a4517ab)
Reviewed-on: http://git-master/r/759239
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jon Mayo <jmayo@nvidia.com>