platform: tegra: add ape power domain wakelock
AGIC is part of APE and AGIC's state is saved/restored
when APE is powered off/on. This is the only way to save
AGIC state as GIC notification is triggered from APE
power off/on.

APE is powered off/on as part of pm_runtime_put/get of APE
devices. This pm runtime work for devices is done through
kworkers. Due to notification audio, ADSPOS is started just
before entering LP0. This starts a race between APE power-off
and system sleep (LP0). If APE is not powered off before
system sleep, AGIC state is not preserved. This causes
mailbox communication (interrupt) issues on next ADSPOS
start and that results in failure to start ADSPOS. And
That leads to adsp audio driver accessing ADSP without it
being functional.

To fix this issue a new wake lock is added for APE. With any
audio playback system does not enter LP0. As this
wake lock is active only when APE is active, this shall
not add any additional wait for system sleep entry (LP0).

Bug 200104336

Change-Id: Ie0d935dc51eedcf75b685782fbd49796ecf82172
Signed-off-by: Nitin Kumbhar <nkumbhar@nvidia.com>
Reviewed-on: http://git-master/r/746480
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sumit Bhattacharya <sumitb@nvidia.com>
Reviewed-by: Viraj Karandikar <vkarandikar@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Prashant Gaikwad <pgaikwad@nvidia.com>