pstore: Removed id from console-ramoops
There are many places in the code, where we
look for file "console-ramoops". We don't
find any match due to entened ID in the name.
Due to which ramoops are skipped from
dropbox and bugreport.

In HE and ST8 we are maintaining only one
session's ramoops and so no need to have ID.

Bug 1646380

Change-Id: Ib4fcd33bc30e144210adb2e13ce486234c06aea4
Signed-off-by: Gagan Grover <ggrover@nvidia.com>
Reviewed-on: http://git-master/r/748586
Reviewed-by: Robert Shih <rshih@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>