firmware: tegra: export bpmp run state
Export an API to return whether bpmp is running the PM firmware or not.
We will use the 'connected' flag to determine whether bpmp firmware is
running.  Technically, it is possible that the firmware is running but
it is not connected (=no IPC established) - but we will ignore this as
an exception.

Change-Id: I685b5ad8f2fffa608b4b331634feb7a67557f952
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/748767
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Vinod Gopalakrishnakurup <vinodg@nvidia.com>
Tested-by: Vinod Gopalakrishnakurup <vinodg@nvidia.com>