ARM: tegra210: Set IO PAD default voltage to 1.8V which only work on 1.8V
On T210, some of IO pads are designed to work on two voltages 1.8V
and 3.3V. The IO-PAD SPI-HV is having the autodetect capability to
detect the voltage level from platform but for other IO PAD, the
capability is removed as part of 20nm design. Hence SW need to
explicitly set the PWR-DET register for the other IO PADS as per
power tree.

All IO pads are not the capable of working on 1.8V and 3.3V
although there is PWR-DET bits on the power detect registers.
These are dummy on nature and just available for backward
compatibility.

The IO pads are:
	SPI-HV:
		Autodetect capability and 1.8/3.3V support.
	GPIO, SDMMC1, SDMMC3, AUDIO_HV:
		Work on 1.8V and 3.3V, SW need to set PWR-DET bit.
	AUDIO, CAM, DBG, DMIC, PEX-CTRL, SPI, UART, SYS, BB:
		Only work on 1.8V.

Configure the IO pad volateg to 1.8V which works on 1.8V only to
match with power tree.

bug 200112284

Change-Id: I8cc3639d6a791f9f261700cb60074ee949987e0f
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/751915