ARM: tegra: hawkeye: use pmc iopower driver for IO POWER control
On T210, there is no auto power detect logic for most of rail and
SW need to override the PMC configuration as per power tree.
Also when IO rail state get change, it is require to control
NO_IOPOWER of PMC for power save and it is done through regulator
notification.

On above context, mode the power detect driver to io-power driver
to control the NO_IOPOWER register.

bug 200112685

Change-Id: I61a6ef61595e8bacb3f4c16464e6a7661c4a2b10
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/753203