ARM: dt: tegra: Add ivm carveout to memory-region property
reg property in ivm carveout is not used. Current entries are dummy
entries to allow reserved-memory framework consider ivm carveout as
a reserved-memory region.

The no-map property is required to make sure memblock_remove() is called
and not memblock_reserve(). memblock_reserve() expects a valid DRAM region
and more over ivm carveout is not intended for kernel's use.

Bug 200027296

Change-Id: I7294391b81d6723dcd1e42ecce069b9608c386eb
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/755697
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Tested-by: Krishna Reddy <vdumpa@nvidia.com>