Revert "net: wireless: bcmdhd2: Add 2nd instance of bcmdhd"
This reverts commit a1bf3e9bd74d8fdce71bb318520b9886bc278604.
Removing the support of the second Wi-Fi chip as of now although
it may be supported in future as this driver is going to be
obsolete for future need. Currently, code is having the coverity
issue and decided to remove the driver from code tree instead of
fixing anything on this as this is obsolete driver.

Bug 200093532

Change-Id: I7b08e39ef07aece555994db1794798e1f5e4be12
Signed-off-by: Raghavender <kraghavender@nvidia.com>
Reviewed-on: http://git-master/r/773853
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>