arm: tegra: hawkeye: keep mipi_1v2 on in shutdown
Currently mipi_1v2 gets disabled by the MAX77620 driver
during shut down. This happens before the DSI driver
gets a chance to properly disable the panel.

Removing the regulator-disable-on-shutdown flag for
the mipi_1v2 and its LDO leaves power on for the DSI
driver. The rail is shut down later in the shut down
sequence.

Bug 1656676

Change-Id: Iab89fb103fea361a6673b83e4469d15bf18e435a
Signed-off-by: Daniel Solomon <daniels@nvidia.com>
Reviewed-on: http://git-master/r/759419
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>