video: tegra: dc: ext: workq becomes kthread
Replace the usage of workq inside the dc flipworker with a
kthread.  This allows us explicit scheduling control over
the kthread running this work, including the ability to make
that thread realtime.  This allows us to tighten the latency
and execution time of the dc flip work.  Since the the display
is updated when the dc work completes instead of on a
vsync pulse in vrr mode the latency and scheduling of the
work becomes much more important.

Bug 1657411

Change-Id: I6ad5052d629fbc3d87475734c731b591d8d0c78a
Reviewed-on: http://git-master/r/760274
Signed-off-by: Jonathan McCaffrey <jmccaffrey@nvidia.com>
Reviewed-on: http://git-master/r/807150
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>