tegra: dc: ext: use work_index 0 for Android
Flip data and flip timestamps are expected to be queued
on a single window's timestamp queue and flip wq. That window
corresponds to work_index, determined at the start of
tegra_dc_ext_flip().

On Android, the work_index mechanism isn't supported - only
Window A's flip and timestamp queues should be used. Currently
this is true for flips, but timestamps can be queued to
different windows' queues depending on work_index.

This change sets work_index for _both_ flip and timestamp
queueing for Android.

Bug 1654135

Change-Id: I3744c151f48f36e8db45ef5edd2c08e8fbbbd13e
Signed-off-by: Daniel Solomon <daniels@nvidia.com>
Reviewed-on: http://git-master/r/761542
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>