hawkeye:dts:add bootloader-orienation-support prop
To support seamless android booting in Hawkeye, we want to load different
image which depends on the device is in landscape or portrait mode.
It is not required in some other devices. Add this attribute into the device
node(/chosen) to ignore the angle value read from accelerometer device.

Bug 200115299

Change-Id: Ifaf368757eae35c72473867ba11c988c12367e44
Signed-off-by: wahsu <wahsu@nvidia.com>
Reviewed-on: http://git-master/r/763407
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>