ARM64: dts: use pmc iopower driver for IO POWER control
On T210, there is no auto power detect logic for most of rail and
SW need to override the PMC configuration as per power tree.
Also when IO rail state get change, it is require to control
NO_IOPOWER register of PMC for power save and it is done through
regulator notification.
On above context, moving the power detect driver to io-power driver
to control the NO_IOPOWER PMC register.

Bug 200119312

Change-Id: I799efb65ab498c9c1881246a04e888adda57d8ca
Signed-off-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
Reviewed-on: http://git-master/r/764744
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>