r8152: use mac address from DTB entry or OTP
Search for MAC address in this order:
1. For built-in device, look in DTB (populated from EEPROM)
2. For add-on device or if DTB does not contain valid entry,
   look in OTP

The factory default setting has the same Nvidia-assigned address
in both OTP and EEPROM.
Either location can be overwritten to override the default address.
The bootloader will add a DTB entry for the address found in EEPROM.

Rationale for this approach:
- Some customers may not have the tools to reprogram OTP, so EEPROM/DTB
needs to have higher precedence than Nvidia OTP setting.
- In case there is an add-on device that also uses this driver, it must
use its own OTP setting, and not steal the address in EEPROM.
- The upstream driver has no knowledge of EEPROM/DTB, so OTP needs to
contain a valid address.

Bug 1655596

Change-Id: Icb62e2d7a31dbd6b350ec5076e83862929b841f3
Signed-off-by: Zheng Liu <zhliu@nvidia.com>
Reviewed-on: http://git-master/r/758238
Reviewed-by: Stephen Warren <swarren@nvidia.com>
Reviewed-by: Matthew Pedro <mapedro@nvidia.com>