arm64: tegra: jetson-e: correct sdmmc4 clock override
pll_c4_out1 is needed to make sdmmc4 to be able to run on 266MHz
for HS533 mode. Remove the incorrect pll source override for sdmmc4.
Also do the following cleanup:
1.tegra210-ers-clk-init.dtsi is already in tegra210-p2530-common.dtsi
so remove it from base dts included files.
2.tegra210-p2530-common.dtsi is already included in jetson-e base dts
file so there is no need to be included in android top dts file.
3.Move therm-est to base dts since it's needed for l4t.
4.Remove duplicated soc_therm and sysedp entries.

Change-Id: I5a88ee9bffc68e764da552b47329e608187aaa86
Signed-off-by: Kerwin Wan <kerwinw@nvidia.com>
Reviewed-on: http://git-master/r/765563
Reviewed-by: Hayden Du <haydend@nvidia.com>