video: tegra: dp: port dp_cust.py to android
dp_cust.py is a syseng script that allows the user to configure
certain DP settings with TegraShell in test mode. This patch ports
this functionality to android as a debugfs node. This debugfs node
is not meant to be used with a real DP sink. See bug below for usage
info.

Bug 1651088

Change-Id: If749fc0cb36d6aff7bbfa5e5472d413231071f2b
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/774160
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Mitch Luban <mluban@nvidia.com>