video: tegra: dc: disable bl in shutdown handler
When we hot-unplug hdmi in console mode, internal display
gets unblanked and backlight brightness is restored to its
current value. This patch moves the call to tegra_dc_disable_bl
to tegra_dc_shutdown so that the backlight brightness isn't
set to zero when internal display is blanked during a hotplug
in console mode.

Bug 200116862
Bug 200107775

Change-Id: Ib9ba448bad86f1235eb04990fbe338a5e4f3ad54
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/779471
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>