video: tegra: nvmap: fix false update of kmap_count
Fixed the incorrect decrement of kmap_count by moving the piece of code
inside proper block. Since decrement of kmap_count is meaningful if and
only if kmap_count is incremented for handle which has an outstanding
kernel map.

In case 'inner' (inner-maintainance) check is 'false', there will be false
decrement of kmap_count. Due to above reason, we receive the following
incorrect output when we read the "iovmm/all_allocation" node.

$cat /d/nvmap/iovmm/all_allocation
    BASE      SIZE USERFLAGS REFS KMAPS UMAPS SHARE    UID
      0       128K  7010001     3     0     1     1 ffffffc0461a6500
      0       128K  7010001     3     0     1     1 ffffffc0461a6700
      0      8832K  c000009     4     0     0     2 ffffffc048125500
>>    0      8192K  c000003     3 4294967295     1     1 ffffffc048294500

This patch fix the above issue by ensuring the proper decrement of kmap_count,
which also satisfies the behavior: REFS >= KMAPS + UMAPS + SHARE.

Bug 1616899

Change-Id: Id36e581b60d39174b9859e51c365cac03b5fd887
Signed-off-by: Amit Sharma (SW-TEGRA) <amisharma@nvidia.com>
Reviewed-on: http://git-master/r/779734
Reviewed-by: Sri Krishna Chowdary <schowdary@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-on: http://git-master/r/780807