ARM: tegra: Hawkeye: make pin pex_l0_rst common
The pin pex_l0_rst is used by the bootloader as an
output GPIO. In order for this pin's pinmux to be
configured correctly by the bootloader, it needs to
be part of the common group. In addition, since it's
an output GPIO, this pin is no longer tristated.

(Previously all this was configured by the kernel, but now
we need this functionality in bootloader as well.)

Bug 1649961

Change-Id: If458767ecb191476d778630f1152e7636cb85a51
Signed-off-by: Daniel Solomon <daniels@nvidia.com>
Reviewed-on: http://git-master/r/783341
(cherry picked from commit c4961707ae44c77ce588568873643a6d38b248ca)
Reviewed-on: http://git-master/r/785015
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>