video: tegra: dc: Expose HDCP through DT
This change exposes the ability to set the HDCP policy through the
device-tree. It moves the HDCP policy into a new device-tree node for
easier maintainace rather than overloading more DC flags. It also adds
a new policy to keep the HDCP always disabled. This is needed for the
recovery kernel where doing HDCP is not possible.

Bug 1694996

Change-Id: Ifd96d2b1bdb3f3f0e963681efd3134719b15e350
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/821655
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>