video: tegra: dc: tag vrr modes using vmode
HDMI VRR is supported only for some resolutions, and the kernel
driver is aware of these restrictions. For VRR monitors, the kernel
driver now duplicates and tags all compatible modes with the
FB_VMODE_VRR flag in the duplicate mode's fb_videomode vmode field.

The same behavior has been added to DSI as well for consistency.

The compatible modes are duplicated before being tagged so as to
not interfere with userspace configurations that do not support
the FB_VMODE_VRR flag.

Bug 1637861

Change-Id: I7acd1f9aa52096b0158ee40a1492bda59eef2bea
Signed-off-by: Daniel Solomon <daniels@nvidia.com>
Reviewed-on: http://git-master/r/833037
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>