usb: phy: Correct usb device speed information
When turn_off_vbus_on_lp0 is set, we disconnect
VBUS before going to LP0 and re-enable it after
LP0 wakeup.

Present code assumes device speed info after LP0
is same as before LP0 which is not true if there
is a swapping of devices when system is in LP0.
Hence set speed as UNKNOWN in such condition.

Bug 200153623

change-Id: I129b88ecc8ad743a0a65341dcb0a7b6001f24ff8
Signed-off-by: Petlozu Pravareshwar <petlozup@nvidia.com>
Reviewed-on: http://git-master/r/835297
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Reviewed-by: Suresh Mangipudi <smangipudi@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>