arm64: tegra: jetson_cv: add dts for boards with super module
On CVB B02 boards, board id of super modulec can not be read
when the debugger board is connected. So we need separate dts
file for boards with super module to unblock SQA stress tests.
This change should be reverted after this issue is fixed.

Bug 200159527

Change-Id: I21bc49f83de143ab3938b13ab672b70eed5e781e
Signed-off-by: Kerwin Wan <kerwinw@nvidia.com>
Reviewed-on: http://git-master/r/927077
GVS: Gerrit_Virtual_Submit