ASoC: Codec: jack det. pin from dts and add workq
Added support to parse the information from i2c device
tree node for using assigned GPIO for jack detection.
The irq is requested based on this.

Defined a local workqueue to handle the jack interrupt
and report to the ASoC jack framework.
(This is specific to k310 version of codec driver, k318
uses the workq provided workqueue framework, which is
not available for k310)

Bug 200148270

Change-Id: I266142b19fa7333cfee88d9f80521cab89ff2b43
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: http://git-master/r/929001
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>