video: tegra: nvmap: Add new NVMAP_PAGES_ZAP_AND_CLEAN option
CUDA needs this zap but not reserve pages in prefetch operations.
That way, if no cpu accesses in between the prefetch operation and
till cuda kernel launch, none of the pages will be marked dirty and
there won't be any zap or cache writeback operation. So, a lot of
overhead can be avoided.

bug 200167711

Change-Id: I3b5d8d12b82ca0b2ebdd94c0f3701f71dd7a7667
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/935713
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>