video: tegra: dp: Correct audio_switch handling
As we saw with HDMI in the past, the correct place
for audio_switch lies with the enable/disable which
is guaranteed to be called always when display turns
on/off. In this bug, when suspend/resume is done,
the audio_switch state doesn't change since it is
only handled from hpd.c (i.e. for hotplug only).

Bug 200161403

Change-Id: I2c3626eb6beed1d47d54d686a691966c2436423e
Signed-off-by: Rahul Mittal <rmittal@nvidia.com>
Reviewed-on: http://git-master/r/933750
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Animesh Kishore <ankishore@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>