tegra-vi2: Fixed register numbers and syncpt issue
New register numbers from TX1 were defined. Additionally, a third channel was defined (new video0 - video2). From a hardware point of view, it is possible to have 6 channels in total (meaning 6 cameras in parallel each using 2 lanes). However, the structure of the driver will change. Thus, only 3 channels were defined (3 camera, each up to 4 lanes), which is comparable to the TK1 and therefore resulted in a minimal change to the driver.

Signed-off-by: Armin Weiss <weii@zhaw.ch>