video: tegra: dc: restore access MUX after regdump
When dumping registers, we switch to active read/write MUX.
But at the end of this function, we don't restore to assembly MUX.
A significant portion of DC accesses are performed with the assumption
that the MUX is set to assembly. This change restores the original MUX
state after dumping registers.

Bug 200105859

Change-Id: I36bdc05064cff06aa1f8fba4e87d5ce4a2d718ee
Signed-off-by: Daniel Solomon <daniels@nvidia.com>
Reviewed-on: http://git-master/r/1011426
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>