ASoC: tegra-alt: jetson-cv - update clock for hra
For HRA use case the codec_rate is set to 96 or 192 KHz.
The mclk/sysclk supplied to codec should be atleaset 256
times of the codec rate.

pll_a_out0 is parent for codec mclk, it is fixed at 49.152MHz.
36.864 MHz was not sufficient for all use cases since 1.5
divider is not supported.
From this required frequencies are derived with interger divider.
Requriements for codec mclk,
    * HRA - 25.576 / 49.152 MHz
    * non-HRA - 12.288 MHz

Also fixed hw_param() failure due to unsupported bit format.
For HRA usecases the format set is 32_LE and same needs to
be acknowledged by codec

Bug 200174763

Change-Id: Ie07615f946c3e828765a8febfa3f341a529deeae
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: http://git-master/r/1012753
GVS: Gerrit_Virtual_Submit
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>