media: i2c: output corrupted frame for cropping
For sensor imx214/imx230 to output corrupted frame
when cropping feature is implemented. Normally, this
frame is not output by the sensor, but that results
in skipping of those frame settings. Hence we need
to output this corrupt frame & later remove it using
bubble capture in SCF.

Bug 1687120

Change-Id: Ifbd88dff66448aca95580502e115cbef703845e4
Signed-off-by: Shubham Chandra <shubhamc@nvidia.com>
Reviewed-on: http://git-master/r/1126892
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ajith Kumar <ajithk@nvidia.com>
Tested-by: Ajith Kumar <ajithk@nvidia.com>
Reviewed-by: Shantanu Nath <snath@nvidia.com>