video: tegra: dc: Enable YUVBYPASS for all modes
YUVBYPASS in its current design was limited for only YUV modes. This change
extends it such at it can be applied to any mode and makes it atomic with
respect to the flip.

Main reason for this patch is for hwcomposer to be able to toggle the
full->limited range conversion for RGB and do it in GLC instead for better
picture quality.

Bug 1750555

Change-Id: I2cd923dbfdd764253d232dcae7eefa310858d61d
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/1132153
(cherry picked from commit fd838ea6331a10a142d455d5748055cc731c5408)
Reviewed-on: http://git-master/r/1142389
GVS: Gerrit_Virtual_Submit
Reviewed-by: Michael I Gold <gold@nvidia.com>
Tested-by: Michael I Gold <gold@nvidia.com>