power: extcon: Avoid race condition
When power supply get registers, it schedule the work queue
for getting property and then return back to caller.

If caller has not initialised all the parameters for reading
property and if initialise after power supply registration
then there is race for work queue get scheduled and use the
parameters which are not initialised and it can cause the
system hang.

Fixing this to initialise parameters for get properties property.

bug 200196661

Change-Id: I8206a61bd3ced9d60135855aa790b10beaa480d3
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1142758
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
Tested-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
GVS: Gerrit_Virtual_Submit